Initial implementation of file-backed durable cuckoo storage engine (#217)
* storage: introduce shm/pmem backed datapool abstraction

This patch creates a datapool abstraction for managing large amounts
of contiguous memory, and adds two implementations:
 - shm, backed by shared memory allocated through cc_alloc
 - pmem, backed by memory-mapped file managed through libpmem

The default implementation, datapool_shm, currently does not differ from
simply managing memory through cc_alloc/cc_free.

The datapool_pmem implementation, which can be enabled through
-DUSE_PMEM=ON, allows for durable (persistent) implementations of the
storage layer as it attempts to recover the previous content
of the managed memory.

* cuckoo: use datapool for cuckoo data storage

This patch uses the datapool interface to allow for durable
operations of the cuckoo storage engine.

To enable durable storage, provide a path to a file in
"cuckoo_datapool" config option. If no file is provided, the
datapool falls back to shared memory.