Reimplement core in rust using async await (#268)
This commit has two parts:
1. It creates a rust implementation of core (rustcore) that uses async-await and tokio 0.2 to abstract over the event loop and allow for multiple types of listeners and protocols to be running at once.
2. It rewrites pingserver-rs to make use of rustcore instead of core.

The general architecture of rustcore is pretty much the same as the architecture of core - there's the same three threads (worker, listener, admin) and the tasks run on each are the same. The big difference is that now tasks can be written using async-await which means that it is easier to see and understand the logic without having to understand how an implementation is threaded through the event loop multiple different times.