WIP: pingserver-rs rewrite
The old pingserver-rs implementation was centered around tokio and
async style programming. When benchmarking, I found the tokio based
implementation had both elevated and inconsitent latency compared
to the C implementation.

I reimplemented the Rust pingserver to use a mio event loop. When
benchmarking, I found that the performance is similar to the C
implementation in terms of latency characteristics and key systems
performance metrics.

NOTE: this is a WIP. Notably, not all configuration options are
being used. There's likely to be some structural changes to the
configuration handling. Additional testing needs to be done to be
sure there is no performance regression since the initial proof of
concept implementation.