pingserver-rs: add rusage metrics (#319)
Adds rusage metrics to pingserver to address #314

Replaces the current metrics library with one being developed for
twemcache-rs. It removes overhead of a hash lookup for each metric
increment by using a fixed indexing from the metric enum into a
vec of channels.

On a test host this is giving ~4-5% gain in peak throughput.