[Segcache] clean up and bug fix  (#326)
* [pre-segcache] preparations for segcache. 1. add a new field in memcache response; 2. change the timer in pelikan_twemcache so that the default timer is memcached_time; 3. fix a bug in pelikan_twemcache where flush_at is not reset during tear and setup. 4. remove inline of _get_value in sarray.c as gcc 7 cannot inline it. 5. add segcache in calculator

* [segcache] the first version before deep clean

* [workflow] add SegcacheExp to GitHub workflow

* add xxhash.h and fix a bug in pthread_setname_np

* [clean] remove constants; separate merge

* [hashtable] clean up Segcache hashtable.

* clean up the code

* the concurrency bug exists in both eviction and merge-based eviction, cleanup to find more clues.

* change xxhash to the lib in ccommon; add more debug information;
fix a data race in item_get where the segment reference count is incrmented out of the hash lock. It is possible that a thread gets the item, but before incrementing refcount, another thread has started and finished evicting the seg;
add a lock in hashtable get, reducing the critical section has a bug, so we currently use locking to protect the hashtable_get, a different solution is to move frequency out of hash table. The same lock adds to hashtable_relink. And I have enabled memset in seg_init to zeroize the seg data.

* disable debug mode

* [bug fix] fix a bug in segmerge where we try to evict empty TTL bucket; fix a bug in evict where min_mature_time is not initialized; clean up code; add lock in hashtable_get to ensure there is no concurrency bug.

* [optimization] remove memset segment data to zero during initialization