rust/session: fix for frequent reallocation on read path
Currently the session uses the minimum buffer size as the desired
read size and reserves free space in the buffer for a full-sized
read on every call. This results in a reallocation on the second
read call for a session.

To avoid frequent reallocation, the session will now only reserve
additional space if there is no available space in the read buffer.