protocol: improve memcache parse performance
Improves memcache request parsing performance by replacing the
use of the pair of slice windows with a manually implementation
which will find specific byte sequences.

This results in better performance by reducing the need to scan
through a very long buffer to find matches for tokenization, as
we can just take the first match as check if it is the sequence
we were looking for or if we need to continue reading.

Adds benchmarks which show an improvement with this change.