bugfix: seg storage CAS may give incorrect result (#337)
The Rust implementation of seg storage appears to have a simple bug
which can result in an incorrect result for a CAS operation.

If the CAS value in the hashtable does not match the provided
value, the code immediately returns that the item exists which is
used to indicate a stale CAS value. However, it does not check if
the item is actually in the table or if this is just a collision.

Changes the code to scan through the items in the bucket even if
the CAS value does not match.