seg: allow disabling compaction for merge eviction
Allows disabling compaction for merge eviction by specifying a
`compact_target` value of 0 in the config.

Compaction is a lossless merge phase which runs on the heel of
item removal if a segment's occupancy is below some watermark. The
`compact_target` setting specifies the minimum number of segments
which should be merged together in this phase. Setting this to 0
should result in disabling compaction.

Changes the handling around calculating the watermark such that
when a `compact_target` of 0 is specified, that compaction is
bypasses as expected.