seg: add RandomFifo eviction strategy
Adds a new eviction strategy that selects a TTL bucket based on the
number of segments per bucket and evicts the head segment for that
bucket.

This strategy preserves the distribution of TTLs in the working set
and prioritizes retaining newer items. Compared to `Random`, this
strategy fits better with the focus of retaining newer items, as it
will always evict the head of a TTL bucket. Compared to `Fifo`, this
strategy should be fairer for workloads with mixed TTLs and varying
write rates per TTL.