update rustcommon to resolve rustsec advisories
Updates rustcommon to resolve RUSTSEC-2020-0071 and
RUSTSEC-2020-0159 advisories both of which have been resolved
by a newer version of rustcommon-time.

This change removes any direct dependency on chrono/time crates
and makes changes to the logging crate to use the latest version
of rustcommon-time. This new version drops local timezone
support, so log output will now always be in UTC regardless of
the system timezone.