Improve CI by properly caching dependencies  (#352)
This commit rewrites the existing CI workflow to take advantage of the features that github actions supports. Namely, caching and the large library of prebuilt actions. This means that most compilations can be much faster since it isn't necessary to recompile all the dependencies over and over again.

Overall, this commit should reduce CI time to ~10 minutes for most changes.