rust: move dependencies on rustcommon into one library
Currently, we face some versioning challenges between the Pelikan
workspace and the fuzz testing suite. The fuzz test suite is
excluded from the Pelikan workspace, so does not share the same
lock file. This means that dependencies are not necessarily in
lockstep with the main code.

This change centralizes the dependencies on rustcommon into the
`common` library within the Pelikan workspace. This allows us to
have a single spot to specify an exact revision of rustcommon for
our dependency.

Includes a bit of refactoring so that the `config` crate depends
on `common` and not the other way around.