rust: add typed values (#384)
Adds typed values to the storage layer. This means that we can
represent numeric values in a native format but allow the protocol 
to present them in a different way. This also allows the protocol to
do the conversion from stringy numbers to typed values, which
can result in space savings and allow for operations such as 
increment / decrement in the future.