segcache-rs: add incr/decr support
Adds support for incr/decr commands to segcache-rs.

In the storage library, we explicitly implement these commands
as wrapping_add() and saturating_sub(). This allows us to choose
the behaviors at higher-levels and potentially make additional
operations available in different protocols.

The implementation in segcache-rs aims to be memcached
compatible and returns "ERROR" when numeric operations are
attempted on non-numeric values.

Replaces #383 and #395