segcache: make item header immutable
Makes the item header immutable in the Rust implementation of
Segcache by removing use of the `deleted` bit in the item header.

Since the Rust implementation of segment storage is not concurrent,
there is no need for the `deleted` bit in the item header. Instead,
we can check the hashtable to confirm if the item is located within
a given segment at a specified offset. By removing the use of the
`deleted` bit, we make the item header immutable. This is
beneficial in that it removes the need to perform random updates
within the segments.