fix reported version numbers (#412)
For the rust servers, the version number reported on the admin port
reflects that of the 'core/server' crate instead of the actual
version number for the binary.

Changes the process and admin builders to allow setting a version
number. Updates the pingserver and segcache binary crates to set
the version number based on their cargo metadata.