Prints the configuration in use for the server and then exits
debatably fixes #331

I could see this needing to be a bit more self-describing in the future, but this will at least make it easy for the enduser to see the precise configuration in use.

I tried doing this with as minimal investment as possible, hence using Serde/TOML, which seems sensible enough given that that's seemingly the configuration language of choice in the Rust part of Pelikan.

fix argument declaration to be more consistent with C version