refactor: move datapool out into separate crate
Moves the datapool out from the seg storage crate and into its own
crate.

Adds a basic header to file-backed datapools and introduces a new
datapool for keeping content in memory with a file available to
save and restore state.

Minor refactoring to accomodate returning a result which allows us
to better surface any issues creating the datapool.