feat: enable metrics descriptions
PR #425 attempted to add metrics description support to Pelikan but
hit some issues due to the way that we have been handling metrics.

This PR adds some new macros to be used in-place of the existing
static_metrics!() macro which allows us to cleanly define metrics
with optional descriptions in the code.

As part of this, metrics are being extensively refactored by moving
the functionality into the 'common' crate.