fix: incorrect logic when removing and replacing items
The logic during removing an item is incorrect. This change fixes
and simplifies the logic when making changes to the item info
entries. Since this implementation is not concurrent, we can
guarantee that the only time we may have a stale item is during an
insert. As the insert logic cleans up the stale item, all other
functions can trust that the first match is the current item.