workspace metadata and update dependencies
Two major things in this PR. First is the changes to the workspace
metadata to make use of the new dependencies and metadata table.
This change allows us to centralize the dependency versions, which
makes updating dependencies much easier. Additionally, we move
several metadata fields (license, edition, etc) and the version
number to be declared at the workspace level. This means that all
crates within Pelikan now share the same version number. The intent
is to make it easier to produce changelogs and cut releases.

The second major thing in this PR is updating the dependencies,
particularly bumping rustcommon up to the latest version which
should bring some performance wins with metrics overhead being
reduced in the latest version.